+++ lib/procon_bypass_man/background.rb
@@ -2,7 +2,6 @@ require "procon_bypass_man/background/jobs/concerns/has_internal_api_setting"
 require "procon_bypass_man/background/jobs/concerns/has_external_api_setting"
 require "procon_bypass_man/background/jobs/concerns/job_performable"
 require "procon_bypass_man/background/job_queue"
-require "procon_bypass_man/background/worker_process"
 require "procon_bypass_man/background/job_performer"
 require "procon_bypass_man/background/jobs/base_job"
 require "procon_bypass_man/background/jobs/report_event_base_job"
